+++ app/logIn/page.js
@@ -1,9 +1,8 @@
-"use server";
 // import { cookies } from "next/headers";
 // import { AUTH_COOKIE_KEY } from "../contants";
 // import { redirect } from "next/navigation";
 import { UserAuthorization } from "../../components/userLogIn/UserAuthorization";
-import { Userlogin } from "../actions";
+// import { Userlogin } from "../actions";
 
 export async function LogIn() {
   // const cookieStore = cookies();
@@ -14,10 +13,10 @@ export async function LogIn() {
   // }
 
   //function to set user info
-  const handleLogIn = async (username, password) => {
-    "use server";
-    await Userlogin(username, password);
-  };
+  // const handleLogIn = async (username, password) => {
+  //   "use server";
+  //   await Userlogin(username, password);
+  // };
 
   return (
     <div className="w-full h-full flex flex-col items-center justify-center py-[20px] px-[30px] mt-8 ">
